add debug component to show route JSON
