+++ deployables/app/app/components/index.ts
@@ -1,4 +1,5 @@
 export { AvatarInput } from './AvatarInput'
 export { ChainSelect } from './ChainSelect'
+export { DebugRouteData } from './DebugRouteData'
 export * from './wallet'
 export { ZodiacMod } from './ZodiacMod'
