+++ deployables/app/app/utils/index.ts
@@ -1,3 +1,4 @@
 export { editRoute } from './editRoute'
 export { jsonRpcProvider } from './jsonRpcProvider'
+export { parseRouteData } from './parseRouteData'
 export { validateAddress } from './validateAddress'
